apps/backend/requirements.txt
fastapi>=0.124.4
uvicorn>=0.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=1.0.10
langchain-google-vertexai>=0.2.0
langgraph>=0.2.3
numpy>=2.0.0
upstash-redis
pydantic-settings>=2.0.0

# GCP/Vertex AI
google-cloud-aiplatform>=1.35.0
google-auth>=2.25.0

# Database - Supabase
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pgvector>=0.2.0

# Supabase Client
supabase>=2.4.0

# Async HTTP client for trend fetching
httpx>=0.27.0

# Reddit API wrapper
praw>=7.7.1

# Twitter API client
tweepy>=4.14.0

# HTML parsing for web scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Embeddings - Local (offline)
sentence-transformers>=3.0.0

# Readability analysis for SEO
textstat>=0.7.3

